requirements.txt
aiohttp==3.8.1
marshmallow==3.14.1
pytest==7.0.0
pytest-asyncio==0.17.2
PyYAML==6.0
gino==1.0.1
asyncpg==0.25.0